Refactor game engine responsibilities into dedicated services
